requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
sqlparse==0.4.4
drf-spectacular==0.22.1
psycopg2-binary==2.9.9

